binance-multi-assets-singal-agent/requirements.txt
fastapi
uvicorn[standard]
python-binance
pandas
pandas-ta
numpy==1.26.4
python-dotenv
websockets
python-multipart 